requirements.txt
# Install with: `pip install -r requirements.txt`

pyzmq == 14.7.0
